feat: expose as experiment function
