+++ parea/__init__.py
@@ -11,7 +11,7 @@
 import sys
 from importlib import metadata as importlib_metadata
 
-from parea.benchmark import run_benchmark
+from parea.experiment.cli import experiment
 from parea.cache import InMemoryCache, RedisCache
 from parea.client import Parea, init
 
@@ -28,8 +28,8 @@ version: str = get_version()
 
 def main():
     args = sys.argv[1:]
-    if args[0] == "benchmark":
-        run_benchmark(args[1:])
+    if args[0] == "experiment":
+        experiment(args[1:])
     else:
         print(f"Unknown command: '{args[0]}'")
